// src/common/enums/file-type.enum.ts
export enum AllowedMimeTypes {
  PDF = 'application/pdf',
  DOCX = 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  DOC = 'application/msword',
  TXT = 'text/plain',
}

export enum AllowedFileExtensions {
  PDF = '.pdf',
  DOCX = '.docx',
  DOC = '.doc',
  TXT = '.txt',
}